inventory_mac: Migrate code out of other places
